Included TinyXml into the project. AntTweakBar(TwOgreGui) now works correctly (Closes issue 1).
New issue
Summary: Add transparent variable references to TwOgreGui
Labels: Type-Enhancement Priority-Medium Milestone-Environment Component-Reliability
Author of TwOgreGui disabled the most important feature of AntTweakBar - the ability to register a variable into the tweaking system so that its value is updated automatically. This flaw needs a workaround or else we might have to deal with a lot of unnecessary boilerplate code.